requirements.txt
decorator
scipy
numpy
matplotlib
pandas
networkx
lxml
pydot
pyyaml
pygraphviz
sortedcontainers
pacsltk
solidpy
sortedcollections
